Move Filter and SortSelect inside features/ui
